+++ includes/utils.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_H
 # define UTILS_H
 
-float idiv(int a, int b);
+double idiv(int a, int b);
 void handle_error(int value);
 void parse_input(char *file, t_scene **scene, t_win *window);
 void	parse_res(char **line, t_scene *scene);
@@ -11,6 +11,6 @@ char	**split_ws(char const *s);
 void	*liberator(char **array);
 char	*strchr_ws(const char *s);
 
-float	ft_strtof(const char *str);
+double	ft_strtof(const char *str);
 
 #endif
