+++ src/fig/square.c
@@ -1,6 +1,6 @@
 #include "minirt.h"
 
-t_square new_square(t_point *pos, t_vec *norm, float side, t_color color)
+t_square new_square(t_point *pos, t_vec *norm, double side, t_color color)
 {
 	t_square result;
