+++ src/geom/vector.c
@@ -1,6 +1,6 @@
 #include "minirt.h"
 
-t_vec	*v_new(float x, float y, float z)
+t_vec	*v_new(double x, double y, double z)
 {
 	t_vec	*vec;
 
@@ -21,9 +21,9 @@ t_vec	*v_sub(t_vec *v1, t_vec *v2)
 	return (result);
 }
 
-float	v_len(t_vec *vec)
+double	v_len(t_vec *vec)
 {
-	float	result;
+	double	result;
 
 	result = sqrt((vec->x * vec->x) + (vec->y * vec->y) + (vec->z * vec->z));
 	return (result);
@@ -31,7 +31,7 @@ float	v_len(t_vec *vec)
 
 void	v_norm(t_vec	*vec)
 {
-	float	len;
+	double	len;
 
 	len = v_len(vec);
 	vec->x /= len;
@@ -39,12 +39,12 @@ void	v_norm(t_vec	*vec)
 	vec->z /= len;
 }
 
-float	v_dot_product(t_vec *v1, t_vec *v2)
+double	v_dot_product(t_vec *v1, t_vec *v2)
 {
 	return ((v1->x * v2->x) + (v1->y * v2->y) + (v1->z * v2->z));
 }
 
-t_vec	*v_mult(t_vec *vec, float mult)
+t_vec	*v_mult(t_vec *vec, double mult)
 {
 	t_vec	*result;
