+++ src/utils/divide.c
@@ -1,6 +1,6 @@
 #include "minirt.h"
 
-float idiv(int a, int b)
+double idiv(int a, int b)
 {
 	return (pow(b, -1) * a);
 }
